COMP5

+++ net/sched/cls_flower_panda_opt.c
@@ -276,7 +276,7 @@ int __ether_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case __cpu_to_be16(ETH_P_IP):
@@ -352,7 +352,7 @@ int __ip_overlay_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case 4:
@@ -413,7 +413,7 @@ int __ipv4_check_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_TCP:
@@ -489,7 +489,7 @@ int __ipv4_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_TCP:
@@ -565,7 +565,7 @@ int __ipv6_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_HOPOPTS:
@@ -653,7 +653,7 @@ int __ipv6_check_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_HOPOPTS:
@@ -741,7 +741,7 @@ int __ipv6_eh_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_HOPOPTS:
@@ -829,7 +829,7 @@ int __ipv6_frag_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case IPPROTO_HOPOPTS:
@@ -917,7 +917,7 @@ int __ppp_node_panda_parse(const struct panda_parser *parser,
 	}
 	
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case __cpu_to_be16(PPP_IP):
@@ -978,7 +978,7 @@ int __pppoe_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case __cpu_to_be16(PPP_IP):
@@ -1039,7 +1039,7 @@ int __e8021AD_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case __cpu_to_be16(ETH_P_IP):
@@ -1115,7 +1115,7 @@ int __e8021Q_node_panda_parse(const struct panda_parser *parser,
 	}
 
 	pr_err(" - offset	: %d", offset);
-	pr_err(" - len		: %d", len);
+	pr_err(" - len		: %d", (int)len);
 
 	switch (type) {
 	case __cpu_to_be16(ETH_P_IP):
